#3 코드리뷰 반영사항
LocationCore 위,경도 Double 타입 변경, WeatherRequestDTO baseURL, serviceKey를 AppConfig로 이동, Location 주석 수정

+++ BeTime/BeTime/Source/Core/LocationCore.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct UserLocation {
   let cityName: String
-  let latitude: Int // 위도
-  let longitude: Int // 경도
+  let latitude: Double // 위도
+  let longitude: Double // 경도
 }
